requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
spotipy==2.24.0
python-dotenv==1.0.1
